% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampleRepo.R
\docType{package}
\name{exampleRepo}
\alias{exampleRepo}
\title{exampleRepo}
\description{
Description of exampleRepo here.
}
\details{
\itemize{
\item Github: https://github.com/JGCRI/exampleRepo
\item Webpage: https://jgcri.github.io/exampleRepo/}
}
