% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampleRepo2.R
\docType{package}
\name{exampleRepo2}
\alias{exampleRepo2}
\title{exampleRepo2}
\description{
This is va package decription
}
\author{
Anna Warmka
\itemize{
\item Github: https://github.com/annawarmka/exampleRepo
\item Webpage: https://annawarmka.github.io/exampleRepo/}
}
